--- a/.changeset/thirty-tools-confess.md
+++ b/.changeset/thirty-tools-confess.md
@@ -1,0 +1,5 @@
+---
+"mobx": minor
+---
+
+Added support for `signal` (AbortSignal) in `autorun` and `reaction` options to dispose those