--- a/docs/reactions.md
+++ b/docs/reactions.md
@@ -364,7 +364,7 @@ Number of milliseconds that can be used to throttle the effect function. If zero
 
 Set a limited amount of time that `when` will wait for. If the deadline passes, `when` will reject / throw.
 
-### `signal` _(autorun, reaction, when)_
+### `signal`
 
 An AbortSignal object instance
 for `when` allows you to abort waiting for the reaction via an AbortController. This will also cause the returned promise to reject with an error "WHEN_ABORTED". This option is ignored when using an effect function, and only applies with the promised based version.