--- a/docs/reactions.md
+++ b/docs/reactions.md
@@ -366,9 +366,8 @@ Set a limited amount of time that `when` will wait for. If the deadline passes, 
 
 ### `signal`
 
-An AbortSignal object instance
-for `when` allows you to abort waiting for the reaction via an AbortController. This will also cause the returned promise to reject with an error "WHEN_ABORTED". This option is ignored when using an effect function, and only applies with the promised based version.
-for `autorun` and `reaction` allow you to dispose the reaction via the AbortController.
+An AbortSignal object instance; can be used as an alternative method for disposal.<br>
+in the promise version of `when` when aborted, the promise rejects with the "WHEN_ABORTED" error.
 
 ### `onError`