--- a/packages/mobx/src/types/generic-abort-signal.ts
+++ b/packages/mobx/src/types/generic-abort-signal.ts
@@ -1,0 +1,7 @@
+// https://github.com/mobxjs/mobx/issues/3582
+export interface GenericAbortSignal {
+    readonly aborted: boolean
+    onabort?: ((...args: any) => any) | null
+    addEventListener?: (...args: any) => any
+    removeEventListener?: (...args: any) => any
+}