@urugator 2023-07-18
What's the use case?