@rluvaton 2023-07-18
When should this be deployed? is there a CD?